requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
psycopg2-binary
python-dotenv
pyautogen
openai
python-multipart